replace the script test with multiple test modules
